requirements.txt
python-fasthtml
uvicorn>=0.29
python-multipart
sqlite-utils
